pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocrclip"
version = "0.1.0"
description = "Simple cross-platform screenshot -> OCR -> clipboard utility"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [ { name = "OCR Clip Authors" } ]
keywords = ["ocr", "screenshot", "clipboard", "pyside6"]

dependencies = [
    "PySide6>=6.6.0",
    "mss>=8.1.0",
    "Pillow>=9.0.0",
    "numpy>=1.24",
    "pynput>=1.7",
    "easyocr>=1.7.2",
    "torchvision>=0.25.0",
    "torch>=2.10.0",
]

[project.optional-dependencies]
full = [
    "easyocr>=1.6",
    "pyinstaller>=5.11",
]
dev = [
    "ruff",
    "black",
    "isort",
    "mypy",
    "pytest",
    "pytest-qt",
]

[project.scripts]
ocrclip = "ocrclip.main:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]
